requirements.txt
python-dotenv
requests
rapidfuzz
azure-communication-email
